Add lib.rs to CLI package to call CLI from other binaries
