+++ splashsurf/src/lib.rs
@@ -1,0 +1,11 @@
+pub mod cli;
+mod convert;
+mod io;
+mod reconstruction;
+#[macro_use]
+mod allocator;
+mod logging;
+#[cfg(test)]
+mod tests;
+
+pub(crate) use register_counting_allocator;
